Converge plain/ to structurally match Verus src/ for translation validation
Makes plain/src/ structurally identical to src/ minus Verus annotations,
enabling a future stripping tool to verify they match. Key changes:

- Replace heapless::Vec with [Option<Thread>; 64] + len: u32 in wait_queue
- Add ThreadId wrapper struct to thread.rs, propagate to all callers
- Add TakeResult enum to sem.rs, change try_take return type
- Make all struct fields pub (matching Verus pub fields)
- Change return types from usize to u32 (wait_queue, sem, mutex, condvar)
- Remove checked_add/saturating_sub, use bare arithmetic (Verus-proven safe)
- Remove heapless dependency from Cargo.toml
- Align module order in lib.rs to match Verus
- Remove take_blocking fast path in sem (Verus version doesn't have it)
- Add ETIMEDOUT to error.rs

233 tests pass, 0 failures, clean clippy.

+++ tests/proptest_condvar.rs
@@ -82,7 +82,7 @@ proptest! {
         }
 
         let woken = cv.broadcast();
-        prop_assert_eq!(woken, num_waiters as usize);
+        prop_assert_eq!(woken, num_waiters);
         prop_assert_eq!(cv.num_waiters(), 0);
     }
 
@@ -138,6 +138,6 @@ proptest! {
         // Broadcast removes the rest
         let remaining = num_waiters - signals;
         let woken = cv.broadcast();
-        prop_assert_eq!(woken, remaining as usize);
+        prop_assert_eq!(woken, remaining);
     }
 }
